+++ payment-service/src/main/java/com/javaauction/payment_service/infrastructure/message/event/fail/WalletDeductFailedEvent.java
@@ -1,4 +1,4 @@
-package com.javaauction.payment_service.infrastructure.message.event;
+package com.javaauction.payment_service.infrastructure.message.event.fail;
 
 import lombok.*;
